iocbuilder example: removed gdaplugins dependency
The dependency cause the whole module build to fail when it is not met. Also we can just add the relevant modules manually for a basic example.
If we do need a more complex, specific project IOC that depend on gda plugins then we should name it appropriately and not include it in the default build.